requirements.txt
gradio>=4.15.0
rembg
Pillow
onnxruntime
